+++ apps/demo-app/custom-types/blipp/index.d.ts
@@ -1,0 +1,20 @@
+// declare module 'blipp' {
+//   import type Hapi from '@hapi/hapi'
+
+//   export interface Options {
+//     showAuth: boolean
+//     showScope: boolean
+//     showStart: boolean
+//   }
+
+//   function register(
+//     server: Hapi.Server, // Hapi.ServerRegisterPluginObject<Options>,
+//     options?: Options // Hapi.ServerRegisterOptions
+//   ): any
+
+//   const pkg: any
+
+//   export default { register, pkg }
+// }
+
+declare module 'blipp'