+++ packages/demo-app/src/main.ts
@@ -1,3 +1,0 @@
-import { provision } from './provision.js'
-
-provision()