+++ packages/hapi-telegram-plugin/__tests__/texts/server-error.test.mjs
@@ -1,0 +1,45 @@
+import {
+  makeHapiServer,
+  requestEventWithData,
+  requestEventWithError
+} from '../../../../scripts/utils.mjs'
+import { serverError } from '../../lib/texts/index.js'
+
+describe('serverError', () => {
+  let server
+  const timeout_ms = 5000
+
+  beforeAll(async () => {
+    server = makeHapiServer()
+  })
+
+  beforeEach(async () => {
+    await server.start()
+  }, timeout_ms)
+
+  afterEach(async () => {
+    await server.stop()
+  }, timeout_ms)
+
+  it('contains the expected substrings', async () => {
+    const method = 'GET'
+    const route_path = '/internal'
+    const url = route_path
+
+    const res = await server.inject({ method, url })
+
+    const request = res.request
+    const event = requestEventWithData({
+      tags: ['foo', 'bar'],
+      data: { abc: 'def', answer: 42 }
+    })
+    const tags = { foo: true, bar: true }
+
+    const s = serverError(request, event, tags)
+
+    expect(s).toContain(method)
+    expect(s).toContain(url)
+    expect(s).toContain(request.info.id)
+    expect(s).toContain(`${event.timestamp}`)
+  })
+})