+++ scripts/utils.mjs
@@ -139,6 +139,18 @@ export const makeDispatch = ({ event, tags, predicate }) => {
   }
 }
 
+export const githubToken = () => {
+  if (process.env.GH_TOKEN) {
+    return process.env.GH_TOKEN
+  }
+
+  if (process.env.GITHUB_TOKEN) {
+    return process.env.GITHUB_TOKEN
+  }
+
+  throw new Error(`neither GH_TOKEN nor GITHUB_TOKEN set in environment`)
+}
+
 export const telegramCredentials = () => {
   let json
   if (process.env.TELEGRAM) {
@@ -173,4 +185,38 @@ export const telegramCredentials = () => {
   }
 
   return { chat_id, token }
+}
+
+/**
+ * see Hapi.RequestEvent
+ */
+export const requestEventWithData = ({
+  tags = ['test'],
+  channel = 'app',
+  data = undefined
+}) => {
+  return {
+    timestamp: new Date().getTime(), // ms
+    tags,
+    channel,
+    // data cannot appear together with error
+    data
+  }
+}
+
+/**
+ * see Hapi.RequestEvent
+ */
+export const requestEventWithError = ({
+  tags = ['test'],
+  channel = 'app',
+  error = new Error('an error occurred')
+}) => {
+  return {
+    timestamp: new Date().getTime(), // ms
+    tags,
+    channel,
+    // the error object related to the event if applicable. Cannot appear together with data.
+    error
+  }
 }